refactored some things to better merge with cleanup
