+++ constants/lighting_constants.py
