+++ hardware/lighting_hardware.py
@@ -1,5 +1,5 @@
 from wpilib._wpilib import AddressableLED
 
-from hardware import port_constants
+from constants import port_constants
 
 led_strip = AddressableLED(port_constants.led_id)
