+++ subsystems/controllers/joystick_drive_controller.py
@@ -1,5 +1,5 @@
 from robot import robot_state
-from hardware import drive_constants
+from constants import drive_constants
 from utils import math_util
 from utils.drive_outputs import DriveOutputs
