+++ subsystems/controllers/move_straight_controller.py
@@ -1,4 +1,4 @@
-from hardware import drive_constants
+from constants import drive_constants
 from utils.drive_outputs import DriveOutputs
