+++ utils/controller_outputs.py
@@ -1,4 +1,4 @@
-from hardware import drive_constants
+from constants import drive_constants
 from utils.gains import Gains
 from ctre import TalonFXControlMode
